refactor: Refactor code, add new tests
