--- tests/widget_movement_tests.rs
@@ -1,1 +1,0 @@
-// TODO: See if we can mock widget movements and test if they work
